requirements.txt
requests
pyyaml
requirements-parser